remove dmx/hal.h
