+++ src/dmx/nvs.h
@@ -9,7 +9,6 @@
  */
 #pragma once
 
-#include "dmx/hal.h"
 #include "dmx_types.h"
 #include "rdm_types.h"
