+++ src/esp_dmx.c
@@ -2,7 +2,1005 @@
 
 #include "dmx/config.h"
 #include "dmx/struct.h"
+#include "dmx/uart.h"
+#include "dmx/timer.h"
+#include "dmx/gpio.h"
+#include "dmx/nvs.h"
 #include "endian.h"
+
+#if ESP_IDF_VERSION_MAJOR >= 5
+#include "esp_mac.h"
+#endif
+
+const char *TAG = "dmx";  // The log tagline for the library
+
+dmx_port_t rdm_binding_port;
+rdm_uid_t rdm_device_uid = {};
+dmx_driver_t *dmx_driver[DMX_NUM_MAX] = {};
+
+static void rdm_default_identify_cb(dmx_port_t dmx_num,
+                                    const rdm_header_t *header, void *context) {
+  if (header->cc == RDM_CC_SET_COMMAND) {
+    const uint8_t *identify =
+        rdm_pd_get(dmx_num, RDM_PID_IDENTIFY_DEVICE, header->sub_device);
+#ifdef ARDUINO
+    printf("RDM identify device is %s\n", *identify ? "on" : "off");
+#else
+    ESP_LOGI(TAG, "RDM identify device is %s", *identify ? "on" : "off");
+#endif
+  }
+}
+
+bool dmx_driver_install(dmx_port_t dmx_num, const dmx_config_t *config,
+                        int intr_flags) {
+  DMX_CHECK(dmx_num < DMX_NUM_MAX, false, "dmx_num error");
+  DMX_CHECK(!dmx_driver_is_installed(dmx_num), false,
+            "driver is already installed");
+  DMX_CHECK(config->dmx_start_address < DMX_PACKET_SIZE_MAX ||
+                config->dmx_start_address == DMX_START_ADDRESS_NONE,
+            false, "dmx_start_address error");
+  DMX_CHECK((config->personality_count == 0 &&
+             config->dmx_start_address == DMX_START_ADDRESS_NONE) ||
+                (config->personality_count > 0 &&
+                 config->personality_count < DMX_PERSONALITY_COUNT_MAX),
+            false, "personality_count error");
+  DMX_CHECK(config->current_personality <= config->personality_count, false,
+            "current_personality error");
+  for (int i = 0; i < config->personality_count; ++i) {
+    DMX_CHECK(config->personalities[i].footprint < DMX_PACKET_SIZE_MAX, false,
+              "footprint error");
+  }
+
+#ifdef DMX_ISR_IN_IRAM
+  // Driver ISR is in IRAM so interrupt flags must include IRAM flag
+  if (!(intr_flags & ESP_INTR_FLAG_IRAM)) {
+    intr_flags |= ESP_INTR_FLAG_IRAM;
+    ESP_LOGI(TAG, "ESP_INTR_FLAG_IRAM flag not set, flag updated");
+  }
+#endif
+
+  // Initialize RDM UID
+  if (rdm_uid_is_null(&rdm_device_uid)) {
+    rdm_device_uid.man_id = RDM_UID_MANUFACTURER_ID;
+    uint32_t dev_id;
+#if RDM_UID_DEVICE_ID == 0xffffffff
+    uint8_t mac[8];
+    esp_efuse_mac_get_default(mac);
+    dev_id = bswap32(*(uint32_t *)(mac + 2));
+#else
+    dev_id = RDM_UID_DEVICE_UID;
+#endif
+    rdm_device_uid.dev_id = dev_id;
+    rdm_binding_port = dmx_num;
+  }
+
+  // Initialize NVS
+  dmx_nvs_init(dmx_num);
+
+  dmx_driver_t *driver;
+
+  // Allocate the DMX driver
+  driver = heap_caps_malloc(sizeof(dmx_driver_t), MALLOC_CAP_8BIT);
+  DMX_CHECK(driver != NULL, false, "DMX driver malloc error");
+  dmx_driver[dmx_num] = driver;
+  driver->mux = NULL;
+  driver->data = NULL;
+  driver->pd = NULL;
+#ifdef DMX_USE_SPINLOCK
+  driver->spinlock = (dmx_spinlock_t)DMX_SPINLOCK_INIT;
+#endif
+
+  // Allocate mutex
+  SemaphoreHandle_t mux = xSemaphoreCreateRecursiveMutex();
+  if (mux == NULL) {
+    dmx_driver_delete(dmx_num);
+    DMX_CHECK(mux != NULL, false, "DMX driver mutex malloc error");
+  }
+
+  // Allocate DMX buffer
+  uint8_t *data = heap_caps_malloc(DMX_PACKET_SIZE, MALLOC_CAP_8BIT);
+  if (data == NULL) {
+    dmx_driver_delete(dmx_num);
+    DMX_CHECK(data != NULL, false, "DMX driver buffer malloc error");
+  }
+  bzero(data, DMX_PACKET_SIZE_MAX);
+
+  // Allocate the RDM parameter buffer
+  bool enable_rdm;
+  size_t pd_size;
+  if (config->pd_size < 53) {
+    // Allocate space for DMX start address and personality info
+    pd_size = sizeof(dmx_driver_personality_t);
+    enable_rdm = false;
+  } else {
+    pd_size = config->pd_size;
+    enable_rdm = true;
+  }
+  uint8_t *pd = heap_caps_malloc(pd_size, MALLOC_CAP_8BIT);
+  if (pd == NULL) {
+    dmx_driver_delete(dmx_num);
+    DMX_CHECK(pd != NULL, false, "DMX driver pd buffer malloc error");
+  }
+
+  // UART configuration
+  driver->dmx_num = dmx_num;
+
+  // Synchronization state
+  driver->mux = mux;
+  driver->task_waiting = NULL;
+
+  // Data buffer
+  driver->head = -1;
+  driver->data = data;
+  driver->tx_size = DMX_PACKET_SIZE_MAX;
+  driver->rx_size = DMX_PACKET_SIZE_MAX;
+
+  // Driver state
+  driver->flags = (DMX_FLAGS_DRIVER_IS_ENABLED | DMX_FLAGS_DRIVER_IS_IDLE);
+  driver->rdm_type = 0;
+  driver->tn = 0;
+  driver->last_slot_ts = 0;
+
+  // DMX configuration
+  driver->break_len = RDM_BREAK_LEN_US;
+  driver->mab_len = RDM_MAB_LEN_US;
+
+  driver->pd_size = pd_size;
+  driver->pd = pd;
+  driver->pd_head = 0;
+
+  // RDM responder configuration
+  driver->rdm_queue_size = 0;
+  driver->rdm_queue_last_sent = 0;  // A queued message has not yet been sent
+  driver->num_rdm_cbs = 0;
+  // The driver->rdm_cbs field is left uninitialized
+
+  // DMX sniffer configuration
+  // The driver->metadata field is left uninitialized
+  driver->metadata_queue = NULL;
+  driver->last_pos_edge_ts = -1;
+  driver->last_neg_edge_ts = -1;
+
+  // Copy the personality table to the driver
+  memcpy(driver->personalities, config->personalities,
+         sizeof(*config->personalities) * config->personality_count);
+
+  // Configure required variables for RDM or DMX-only
+  if (enable_rdm) {
+    uint16_t footprint;
+    if (config->personality_count > 0 && config->current_personality > 0) {
+      const uint8_t personality_idx = config->current_personality - 1;
+      footprint = driver->personalities[personality_idx].footprint;
+    } else {
+      footprint = 0;
+    }
+    rdm_device_info_t device_info = {
+        .model_id = config->model_id,
+        .product_category = config->product_category,
+        .software_version_id = config->software_version_id,
+        .footprint = footprint,
+        .current_personality = config->current_personality,
+        .personality_count = config->personality_count,
+        .dmx_start_address = config->dmx_start_address,
+        .sub_device_count = 0,  // Sub-devices must be registered
+        .sensor_count = 0,      // Sensors must be registered
+    };
+    rdm_register_supported_parameters(dmx_num, NULL, NULL);
+    rdm_register_disc_unique_branch(dmx_num, NULL, NULL);
+    rdm_register_disc_un_mute(dmx_num, NULL, NULL);
+    rdm_register_disc_mute(dmx_num, NULL, NULL);
+    rdm_register_device_info(dmx_num, &device_info, NULL, NULL);
+    rdm_register_device_label(dmx_num, config->device_label, NULL, NULL);
+    rdm_register_software_version_label(dmx_num, config->software_version_label,
+                                        NULL, NULL);
+    rdm_register_identify_device(dmx_num, rdm_default_identify_cb, NULL);
+    rdm_register_dmx_personality(dmx_num, NULL, NULL);
+    rdm_register_dmx_personality_description(dmx_num, NULL, NULL);
+    rdm_register_parameter_description(dmx_num, NULL, NULL);
+
+    if (device_info.dmx_start_address != DMX_START_ADDRESS_NONE) {
+      rdm_register_dmx_start_address(dmx_num, NULL, NULL);
+    }
+  } else {
+    dmx_driver_personality_t *dmx = rdm_pd_alloc(dmx_num, pd_size);
+    assert(dmx != NULL);
+
+    // Load the DMX start address from NVS
+    uint16_t dmx_start_address;
+    if (config->dmx_start_address == 0) {
+      size_t size = sizeof(dmx_start_address);
+      if (!dmx_nvs_get(dmx_num, RDM_PID_DMX_START_ADDRESS, RDM_DS_UNSIGNED_WORD,
+                       &dmx_start_address, &size)) {
+        dmx_start_address = 1;
+      }
+    } else {
+      dmx_start_address = config->dmx_start_address;
+    }
+
+    // Load the current DMX personality from NVS
+    uint8_t current_personality;
+    if (config->current_personality == 0 &&
+        dmx_start_address != DMX_START_ADDRESS_NONE) {
+      rdm_dmx_personality_t personality;
+      size_t size = sizeof(personality);
+      if (!dmx_nvs_get(dmx_num, RDM_PID_DMX_PERSONALITY, RDM_DS_BIT_FIELD,
+                       &personality, &size) ||
+          personality.personality_count != config->personality_count) {
+        current_personality = 1;
+      } else {
+        current_personality = personality.current_personality;
+      }
+    } else {
+      current_personality = config->current_personality;
+    }
+
+    dmx->dmx_start_address = dmx_start_address;
+    dmx->current_personality = current_personality;
+    dmx->personality_count = config->personality_count;
+  }
+
+  // Initialize the UART peripheral
+  driver->uart = dmx_uart_init(dmx_num, driver, intr_flags);
+  if (driver->uart == NULL) {
+    dmx_driver_delete(dmx_num);
+    DMX_CHECK(driver->uart != NULL, false, "UART init error");
+  }
+
+  // Initialize the timer peripheral
+  driver->timer = dmx_timer_init(dmx_num, driver, intr_flags);
+  if (driver->timer == NULL) {
+    dmx_driver_delete(dmx_num);
+    DMX_CHECK(driver->timer != NULL, false, "timer init error");
+  }
+
+  // Enable reading on the DMX port
+  taskENTER_CRITICAL(DMX_SPINLOCK(dmx_num));
+  xTaskNotifyStateClear(xTaskGetCurrentTaskHandle());
+  dmx_uart_enable_interrupt(driver->uart, DMX_INTR_RX_ALL);
+  dmx_uart_set_rts(driver->uart, 1);
+  taskEXIT_CRITICAL(DMX_SPINLOCK(dmx_num));
+
+  // Give the mutex and return
+  xSemaphoreGiveRecursive(driver->mux);
+  return true;
+}
+
+bool dmx_driver_delete(dmx_port_t dmx_num) {
+  DMX_CHECK(dmx_num < DMX_NUM_MAX, false, "dmx_num error");
+  DMX_CHECK(dmx_driver_is_installed(dmx_num), false, "driver is not installed");
+
+  dmx_driver_t *const driver = dmx_driver[dmx_num];
+
+  // Free driver mutex
+  if (!xSemaphoreTakeRecursive(driver->mux, 0)) {
+    return false;
+  }
+  xSemaphoreGiveRecursive(driver->mux);
+  vSemaphoreDelete(driver->mux);
+
+  // Uninstall sniffer ISR
+  if (dmx_sniffer_is_enabled(dmx_num)) {
+    dmx_sniffer_disable(dmx_num);
+  }
+
+  // Free driver data buffer
+  if (driver->data != NULL) {
+    heap_caps_free(driver->data);
+  }
+
+  // Free RDM parameter data buffer
+  if (driver->pd != NULL) {
+    heap_caps_free(driver->pd);
+  }
+
+  // Free hardware timer ISR
+  dmx_timer_deinit(driver->timer);
+
+  // Disable UART module
+  dmx_uart_deinit(driver->uart);
+
+  // Free driver
+  heap_caps_free(driver);
+  dmx_driver[dmx_num] = NULL;
+
+  return true;
+}
+
+bool dmx_driver_disable(dmx_port_t dmx_num) {
+  DMX_CHECK(dmx_num < DMX_NUM_MAX, false, "dmx_num error");
+  DMX_CHECK(dmx_driver_is_installed(dmx_num), false, "driver is not installed");
+  DMX_CHECK(dmx_driver_is_enabled(dmx_num), false,
+            "driver is already disabled");
+
+  dmx_driver_t *const driver = dmx_driver[dmx_num];
+
+  // Disable receive interrupts
+  bool ret = false;
+  taskENTER_CRITICAL(DMX_SPINLOCK(dmx_num));
+  if (!(driver->flags & DMX_FLAGS_DRIVER_IS_SENDING)) {
+    dmx_uart_disable_interrupt(driver->uart, DMX_INTR_RX_ALL);
+    dmx_uart_clear_interrupt(driver->uart, DMX_INTR_RX_ALL);
+    driver->flags &= ~DMX_FLAGS_DRIVER_IS_ENABLED;
+    ret = true;
+  }
+  taskEXIT_CRITICAL(DMX_SPINLOCK(dmx_num));
+
+  return ret;
+}
+
+bool dmx_driver_enable(dmx_port_t dmx_num) {
+  DMX_CHECK(dmx_num < DMX_NUM_MAX, false, "dmx_num error");
+  DMX_CHECK(dmx_driver_is_installed(dmx_num), false, "driver is not installed");
+  DMX_CHECK(!dmx_driver_is_enabled(dmx_num), false,
+            "driver is already enabled");
+
+  dmx_driver_t *const driver = dmx_driver[dmx_num];
+
+  // Initialize driver flags and reenable interrupts
+  taskENTER_CRITICAL(DMX_SPINLOCK(dmx_num));
+  driver->head = -1;  // Wait for DMX break before reading data
+  driver->flags |= (DMX_FLAGS_DRIVER_IS_ENABLED | DMX_FLAGS_DRIVER_IS_IDLE);
+  driver->flags &= ~(DMX_FLAGS_DRIVER_IS_IN_BREAK | DMX_FLAGS_DRIVER_HAS_DATA);
+  dmx_uart_rxfifo_reset(driver->uart);
+  dmx_uart_txfifo_reset(driver->uart);
+  dmx_uart_enable_interrupt(driver->uart, DMX_INTR_RX_ALL);
+  dmx_uart_clear_interrupt(driver->uart, DMX_INTR_RX_ALL);
+  taskEXIT_CRITICAL(DMX_SPINLOCK(dmx_num));
+
+  return true;
+}
+
+bool dmx_set_pin(dmx_port_t dmx_num, int tx_pin, int rx_pin, int rts_pin) {
+  DMX_CHECK(dmx_num < DMX_NUM_MAX, false, "dmx_num error");
+  DMX_CHECK(dmx_tx_pin_is_valid(tx_pin), false, "tx_pin error");
+  DMX_CHECK(dmx_rx_pin_is_valid(rx_pin), false, "rx_pin error");
+  DMX_CHECK(dmx_rts_pin_is_valid(rts_pin), false, "rts_pin error");
+  DMX_CHECK(dmx_driver_is_installed(dmx_num), false, "driver is not installed");
+
+  return dmx_uart_set_pin(dmx_driver[dmx_num]->uart, tx_pin, rx_pin, rts_pin);
+}
+
+uint32_t dmx_set_baud_rate(dmx_port_t dmx_num, uint32_t baud_rate) {
+  DMX_CHECK(dmx_num < DMX_NUM_MAX, 0, "dmx_num error");
+
+  // Clamp the baud rate to within DMX specification
+  if (baud_rate < DMX_BAUD_RATE_MIN) {
+    baud_rate = DMX_BAUD_RATE_MIN;
+  } else if (baud_rate > DMX_BAUD_RATE_MAX) {
+    baud_rate = DMX_BAUD_RATE_MAX;
+  }
+
+  taskENTER_CRITICAL(DMX_SPINLOCK(dmx_num));
+  dmx_uart_set_baud_rate(dmx_driver[dmx_num]->uart, baud_rate);
+  taskEXIT_CRITICAL(DMX_SPINLOCK(dmx_num));
+
+  return baud_rate;
+}
+
+uint32_t dmx_get_baud_rate(dmx_port_t dmx_num) {
+  DMX_CHECK(dmx_num < DMX_NUM_MAX, 0, "dmx_num error");
+
+  uint32_t baud_rate;
+  taskENTER_CRITICAL(DMX_SPINLOCK(dmx_num));
+  baud_rate = dmx_uart_get_baud_rate(dmx_driver[dmx_num]->uart);
+  taskEXIT_CRITICAL(DMX_SPINLOCK(dmx_num));
+
+  return baud_rate;
+}
+
+bool dmx_set_start_address(dmx_port_t dmx_num, uint16_t dmx_start_address) {
+  DMX_CHECK(dmx_num < DMX_NUM_MAX, 0, "dmx_num error");
+  DMX_CHECK(dmx_driver_is_installed(dmx_num), 0, "driver is not installed");
+  DMX_CHECK(dmx_start_address > 0 && dmx_start_address < DMX_PACKET_SIZE_MAX,
+            false, "dmx_start_address error");
+  DMX_CHECK(dmx_get_start_address(dmx_num) != DMX_START_ADDRESS_NONE, false,
+            "cannot set DMX start address");
+
+  // TODO: make a function to check if RDM is enabled on the driver
+  const bool rdm_is_enabled = (dmx_driver[dmx_num]->pd_size >= 53);
+
+  if (rdm_is_enabled) {
+    rdm_pd_set(dmx_num, RDM_PID_DMX_START_ADDRESS, RDM_SUB_DEVICE_ROOT,
+                      &dmx_start_address, sizeof(uint16_t), true);
+  } else {
+    taskENTER_CRITICAL(DMX_SPINLOCK(dmx_num));
+    dmx_driver_personality_t *personality = (void *)dmx_driver[dmx_num]->pd;
+    personality->dmx_start_address = dmx_start_address;
+    taskEXIT_CRITICAL(DMX_SPINLOCK(dmx_num));
+  }
+
+  return true;
+}
+
+size_t dmx_write_offset(dmx_port_t dmx_num, size_t offset, const void *source,
+                        size_t size) {
+  DMX_CHECK(dmx_num < DMX_NUM_MAX, 0, "dmx_num error");
+  DMX_CHECK(offset < DMX_PACKET_SIZE_MAX, 0, "offset error");
+  DMX_CHECK(source, 0, "source is null");
+  DMX_CHECK(dmx_driver_is_installed(dmx_num), 0, "driver is not installed");
+
+  // Clamp size to the maximum DMX packet size
+  if (size + offset > DMX_PACKET_SIZE_MAX) {
+    size = DMX_PACKET_SIZE_MAX - offset;
+  } else if (size == 0) {
+    return 0;
+  }
+
+  dmx_driver_t *const driver = dmx_driver[dmx_num];
+
+  taskENTER_CRITICAL(DMX_SPINLOCK(dmx_num));
+  if ((driver->flags & DMX_FLAGS_DRIVER_IS_SENDING) && driver->rdm_type != 0) {
+    // Do not allow asynchronous writes when sending an RDM packet
+    taskEXIT_CRITICAL(DMX_SPINLOCK(dmx_num));
+    return 0;
+  } else if (dmx_uart_get_rts(driver->uart) == 1) {
+    // Flip the bus to stop writes from being overwritten by incoming data
+    dmx_uart_set_rts(driver->uart, 0);
+  }
+  driver->tx_size = offset + size;  // Update driver transmit size
+
+  // Copy data from the source to the driver buffer asynchronously
+  memcpy(driver->data + offset, source, size);
+
+  taskEXIT_CRITICAL(DMX_SPINLOCK(dmx_num));
+
+  return size;
+}
+
+size_t dmx_write_rdm(dmx_port_t dmx_num, rdm_header_t *header, const void *pd) {
+  DMX_CHECK(dmx_num < DMX_NUM_MAX, 0, "dmx_num error");
+  DMX_CHECK(header != NULL || pd != NULL, 0,
+            "header is null and pd does not contain a UID");
+  DMX_CHECK(dmx_driver_is_installed(dmx_num), 0, "driver is not installed");
+
+  size_t written = 0;
+
+  dmx_driver_t *const driver = dmx_driver[dmx_num];
+
+  // Get pointers to driver data buffer locations and declare checksum
+  uint16_t checksum = 0;
+  uint8_t *header_ptr = driver->data;
+  void *pd_ptr = header_ptr + 24;
+
+  // RDM writes must be synchronous to prevent data corruption
+  taskENTER_CRITICAL(DMX_SPINLOCK(dmx_num));
+  if (driver->flags & DMX_FLAGS_DRIVER_IS_SENDING) {
+    taskEXIT_CRITICAL(DMX_SPINLOCK(dmx_num));
+    return written;
+  } else if (dmx_uart_get_rts(driver->uart) == 1) {
+    dmx_uart_set_rts(driver->uart, 0);  // Stops writes from being overwritten
+  }
+
+  if (header != NULL && !(header->cc == RDM_CC_DISC_COMMAND_RESPONSE &&
+                          header->pid == RDM_PID_DISC_UNIQUE_BRANCH)) {
+    // Copy the header, pd, message_len, and pdl into the driver
+    const size_t copy_size = header->pdl <= 231 ? header->pdl : 231;
+    header->message_len = copy_size + 24;
+    rdm_pd_emplace(header_ptr, "#cc01hbuubbbwbwb", header, sizeof(*header),
+                   false);
+    memcpy(pd_ptr, pd, copy_size);
+
+    // Calculate and copy the checksum
+    checksum = RDM_SC + RDM_SUB_SC;
+    for (int i = 2; i < header->message_len; ++i) {
+      checksum += header_ptr[i];
+    }
+    *(uint16_t *)(header_ptr + header->message_len) = bswap16(checksum);
+
+    // Update written size
+    written = header->message_len + 2;
+  } else {
+    // Encode the preamble bytes
+    const size_t preamble_len = 7;
+    for (int i = 0; i < preamble_len; ++i) {
+      header_ptr[i] = RDM_PREAMBLE;
+    }
+    header_ptr[preamble_len] = RDM_DELIMITER;
+    header_ptr += preamble_len + 1;
+
+    // Encode the UID and calculate the checksum
+    uint8_t uid[6];
+    if (header == NULL) {
+      memcpy(uid, pd, sizeof(rdm_uid_t));
+    } else {
+      rdm_uidcpy(uid, &header->src_uid);
+    }
+    for (int i = 0, j = 0; j < sizeof(rdm_uid_t); i += 2, ++j) {
+      header_ptr[i] = uid[j] | 0xaa;
+      header_ptr[i + 1] = uid[j] | 0x55;
+      checksum += uid[j] + (0xaa | 0x55);
+    }
+    header_ptr += sizeof(rdm_uid_t) * 2;
+
+    // Encode the checksum
+    header_ptr[0] = (uint8_t)(checksum >> 8) | 0xaa;
+    header_ptr[1] = (uint8_t)(checksum >> 8) | 0x55;
+    header_ptr[2] = (uint8_t)(checksum) | 0xaa;
+    header_ptr[3] = (uint8_t)(checksum) | 0x55;
+
+    // Update written size
+    written = preamble_len + 1 + 16;
+  }
+
+  // Update driver transmission size
+  driver->tx_size = written;
+  taskEXIT_CRITICAL(DMX_SPINLOCK(dmx_num));
+
+  return written;
+}
+
+size_t dmx_receive(dmx_port_t dmx_num, dmx_packet_t *packet,
+                   TickType_t wait_ticks) {
+  DMX_CHECK(dmx_num < DMX_NUM_MAX, 0, "dmx_num error");
+  DMX_CHECK(dmx_driver_is_installed(dmx_num), 0, "driver is not installed");
+  DMX_CHECK(dmx_driver_is_enabled(dmx_num), 0, "driver is not enabled");
+
+  dmx_driver_t *const driver = dmx_driver[dmx_num];
+
+  // Set default return value and default values for output argument
+  dmx_err_t err = DMX_OK;
+  uint32_t packet_size = 0;
+  if (packet != NULL) {
+    packet->err = DMX_ERR_TIMEOUT;
+    packet->sc = -1;
+    packet->size = 0;
+    packet->is_rdm = 0;
+  }
+
+  // Block until mutex is taken and driver is idle, or until a timeout
+  TimeOut_t timeout;
+  vTaskSetTimeOutState(&timeout);
+  if (!xSemaphoreTakeRecursive(driver->mux, wait_ticks) ||
+      (wait_ticks && xTaskCheckForTimeOut(&timeout, &wait_ticks))) {
+    return packet_size;
+  } else if (!dmx_wait_sent(dmx_num, wait_ticks) ||
+             (wait_ticks && xTaskCheckForTimeOut(&timeout, &wait_ticks))) {
+    xSemaphoreGiveRecursive(driver->mux);
+    return packet_size;
+  }
+
+  // Set the RTS pin to enable reading from the DMX bus
+  if (dmx_uart_get_rts(driver->uart) == 0) {
+    taskENTER_CRITICAL(DMX_SPINLOCK(dmx_num));
+    xTaskNotifyStateClear(xTaskGetCurrentTaskHandle());
+    driver->head = -1;  // Wait for DMX break before reading data
+    driver->flags &= ~DMX_FLAGS_DRIVER_HAS_DATA;
+    dmx_uart_set_rts(driver->uart, 1);
+    taskEXIT_CRITICAL(DMX_SPINLOCK(dmx_num));
+  }
+
+  // Wait for new DMX packet to be received
+  taskENTER_CRITICAL(DMX_SPINLOCK(dmx_num));
+  int driver_flags = driver->flags;
+  taskEXIT_CRITICAL(DMX_SPINLOCK(dmx_num));
+  if (!(driver_flags & DMX_FLAGS_DRIVER_HAS_DATA) && wait_ticks > 0) {
+    // Set task waiting and get additional DMX driver flags
+    taskENTER_CRITICAL(DMX_SPINLOCK(dmx_num));
+    const int rdm_type = driver->rdm_type;
+    driver->task_waiting = xTaskGetCurrentTaskHandle();
+    taskEXIT_CRITICAL(DMX_SPINLOCK(dmx_num));
+
+    // Check for early timeout according to RDM specification
+    const int RDM_EARLY_TIMEOUT =
+        (DMX_FLAGS_RDM_IS_REQUEST | DMX_FLAGS_RDM_IS_DISC_UNIQUE_BRANCH);
+    if ((driver_flags & DMX_FLAGS_DRIVER_SENT_LAST) &&
+        (rdm_type & RDM_EARLY_TIMEOUT) == RDM_EARLY_TIMEOUT) {
+      taskENTER_CRITICAL(DMX_SPINLOCK(dmx_num));
+      const int64_t last_timestamp = driver->last_slot_ts;
+      taskEXIT_CRITICAL(DMX_SPINLOCK(dmx_num));
+
+      // Guard against setting hardware alarm durations with negative values
+      int64_t elapsed = dmx_timer_get_micros_since_boot() - last_timestamp;
+      if (elapsed >= RDM_PACKET_SPACING_CONTROLLER_NO_RESPONSE) {
+        taskENTER_CRITICAL(DMX_SPINLOCK(dmx_num));
+        driver->task_waiting = NULL;
+        taskEXIT_CRITICAL(DMX_SPINLOCK(dmx_num));
+        xSemaphoreGiveRecursive(driver->mux);
+        return packet_size;
+      }
+
+      // Set an early timeout with the hardware timer
+      taskENTER_CRITICAL(DMX_SPINLOCK(dmx_num));
+      dmx_timer_set_counter(driver->timer, elapsed);
+      dmx_timer_set_alarm(driver->timer,
+                          RDM_PACKET_SPACING_CONTROLLER_NO_RESPONSE, false);
+      dmx_timer_start(driver->timer);
+      taskEXIT_CRITICAL(DMX_SPINLOCK(dmx_num));
+    }
+
+    // Wait for a task notification
+    const bool notified = xTaskNotifyWait(0, -1, (uint32_t *)&err, wait_ticks);
+    taskENTER_CRITICAL(DMX_SPINLOCK(dmx_num));
+    dmx_timer_stop(driver->timer);
+    driver->task_waiting = NULL;
+    taskEXIT_CRITICAL(DMX_SPINLOCK(dmx_num));
+    if (!notified) {
+      xTaskNotifyStateClear(xTaskGetCurrentTaskHandle());
+      xSemaphoreGiveRecursive(driver->mux);
+      return packet_size;
+    }
+  } else if (!(driver_flags & DMX_FLAGS_DRIVER_HAS_DATA)) {
+    // Fail early if there is no data available and this function cannot block
+    xSemaphoreGiveRecursive(driver->mux);
+    return packet_size;
+  }
+
+  // Parse DMX data packet
+  taskENTER_CRITICAL(DMX_SPINLOCK(dmx_num));
+  driver->flags &= ~DMX_FLAGS_DRIVER_HAS_DATA;
+  packet_size = driver->head;
+  taskEXIT_CRITICAL(DMX_SPINLOCK(dmx_num));
+  if (packet_size == -1) {
+    packet_size = 0;
+  }
+  if (packet != NULL) {
+    taskENTER_CRITICAL(DMX_SPINLOCK(dmx_num));
+    packet->sc = packet_size > 0 ? driver->data[0] : -1;
+    driver->flags &= ~DMX_FLAGS_DRIVER_HAS_DATA;
+    taskEXIT_CRITICAL(DMX_SPINLOCK(dmx_num));
+    packet->err = err;
+    packet->size = packet_size;
+    packet->is_rdm = 0;
+  }
+
+  // Return early if the no data was received
+  if (packet_size == 0) {
+    xSemaphoreGiveRecursive(driver->mux);
+    return packet_size;
+  }
+
+  // Return early if the packet is neither RDM nor an RDM request
+  rdm_header_t header;
+  if (!dmx_read_rdm(dmx_num, &header, NULL, 0) ||
+      (header.cc != RDM_CC_DISC_COMMAND && header.cc != RDM_CC_GET_COMMAND &&
+       header.cc != RDM_CC_SET_COMMAND)) {
+    xSemaphoreGiveRecursive(driver->mux);
+    return packet_size;
+  }
+  if (packet != NULL) {
+    packet->is_rdm = header.pid;
+  }
+
+  // Ignore the packet if it does not target this device
+  rdm_uid_t my_uid;
+  rdm_uid_get(dmx_num, &my_uid);
+  if (!rdm_uid_is_target(&my_uid, &header.dest_uid)) {
+    // The packet should be ignored
+    xSemaphoreGiveRecursive(driver->mux);
+    return packet_size;
+  }
+
+  // Prepare the response packet parameter data and find the correct callback
+  rdm_response_type_t response_type;
+  uint8_t pdl_out;
+  uint8_t pd[231];
+  int cb_num = 0;
+  for (; cb_num < driver->num_rdm_cbs; ++cb_num) {
+    if (driver->rdm_cbs[cb_num].desc.pid == header.pid) {
+      break;
+    }
+  }
+  const rdm_pid_description_t *desc;
+  void *param;
+  if (cb_num < driver->num_rdm_cbs) {
+    desc = &driver->rdm_cbs[cb_num].desc;
+    param = driver->rdm_cbs[cb_num].param;
+  } else {
+    desc = NULL;
+    param = NULL;
+  }
+
+  // Determine how this device should respond to the request
+  if (header.pdl > sizeof(pd) || header.port_id == 0 ||
+      rdm_uid_is_broadcast(&header.src_uid)) {
+    // The packet format is invalid
+    response_type = RDM_RESPONSE_TYPE_NACK_REASON;
+    pdl_out = rdm_pd_emplace_word(pd, RDM_NR_FORMAT_ERROR);
+  } else if (cb_num == driver->num_rdm_cbs) {
+    // The requested PID is unknown
+    response_type = RDM_RESPONSE_TYPE_NACK_REASON;
+    pdl_out = rdm_pd_emplace_word(pd, RDM_NR_UNKNOWN_PID);
+  } else if ((header.cc == RDM_CC_DISC_COMMAND && desc->cc != RDM_CC_DISC) ||
+             (header.cc == RDM_CC_GET_COMMAND && !(desc->cc & RDM_CC_GET)) ||
+             (header.cc == RDM_CC_SET_COMMAND && !(desc->cc & RDM_CC_SET))) {
+    // The PID does not support the request command class
+    response_type = RDM_RESPONSE_TYPE_NACK_REASON;
+    pdl_out = rdm_pd_emplace_word(pd, RDM_NR_UNSUPPORTED_COMMAND_CLASS);
+  } else if ((header.sub_device > 512 &&
+              header.sub_device != RDM_SUB_DEVICE_ALL) ||
+             (header.sub_device == RDM_SUB_DEVICE_ALL &&
+              header.cc == RDM_CC_GET_COMMAND)) {
+    // The sub-device is out of range
+    response_type = RDM_RESPONSE_TYPE_NACK_REASON;
+    pdl_out = rdm_pd_emplace_word(pd, RDM_NR_SUB_DEVICE_OUT_OF_RANGE);
+  } else {
+    // Call the appropriate driver-side RDM callback to process the request
+    pdl_out = 0;
+    dmx_read_rdm(dmx_num, NULL, pd, sizeof(pd));
+    const char *param_str = driver->rdm_cbs[cb_num].param_str;
+    response_type = driver->rdm_cbs[cb_num].driver_cb(
+        dmx_num, &header, pd, &pdl_out, param_str);
+
+    // Verify that the driver-side callback returned correctly
+    if (pdl_out > sizeof(pd)) {
+      DMX_WARN("PID 0x%04x pdl is too large", header.pid);
+      taskENTER_CRITICAL(DMX_SPINLOCK(dmx_num));
+      driver->flags |= DMX_FLAGS_DRIVER_BOOT_LOADER;
+      taskEXIT_CRITICAL(DMX_SPINLOCK(dmx_num));
+      response_type = RDM_RESPONSE_TYPE_NACK_REASON;
+      pdl_out = rdm_pd_emplace_word(pd, RDM_NR_HARDWARE_FAULT);
+    } else if ((response_type != RDM_RESPONSE_TYPE_NONE &&
+                response_type != RDM_RESPONSE_TYPE_ACK &&
+                response_type != RDM_RESPONSE_TYPE_ACK_TIMER &&
+                response_type != RDM_RESPONSE_TYPE_NACK_REASON &&
+                response_type != RDM_RESPONSE_TYPE_ACK_OVERFLOW) ||
+               (response_type == RDM_RESPONSE_TYPE_NONE &&
+                (header.pid != RDM_PID_DISC_UNIQUE_BRANCH ||
+                 !rdm_uid_is_broadcast(&header.dest_uid))) ||
+               ((response_type != RDM_RESPONSE_TYPE_ACK &&
+                 response_type != RDM_RESPONSE_TYPE_NONE) &&
+                header.cc == RDM_CC_DISC_COMMAND)) {
+      DMX_WARN("PID 0x%04x returned invalid response type", header.pid);
+      taskENTER_CRITICAL(DMX_SPINLOCK(dmx_num));
+      driver->flags |= DMX_FLAGS_DRIVER_BOOT_LOADER;
+      taskEXIT_CRITICAL(DMX_SPINLOCK(dmx_num));
+      response_type = RDM_RESPONSE_TYPE_NACK_REASON;
+      pdl_out = rdm_pd_emplace_word(pd, RDM_NR_HARDWARE_FAULT);
+    }
+  }
+
+  // Don't respond to non-discovery broadcasts nor send NACK to DISC packets
+  if ((rdm_uid_is_broadcast(&header.dest_uid) &&
+       header.pid != RDM_PID_DISC_UNIQUE_BRANCH) ||
+      (response_type != RDM_RESPONSE_TYPE_ACK &&
+       header.cc == RDM_CC_DISC_COMMAND)) {
+    response_type = RDM_RESPONSE_TYPE_NONE;
+  }
+
+  // Rewrite the header for the response packet
+  header.message_len = 24 + pdl_out;  // Set for user callback
+  header.dest_uid = header.src_uid;
+  header.src_uid = my_uid;
+  header.response_type = response_type;
+  taskENTER_CRITICAL(DMX_SPINLOCK(dmx_num));
+  header.message_count = driver->rdm_queue_size;
+  taskEXIT_CRITICAL(DMX_SPINLOCK(dmx_num));
+  header.cc += 1;  // Set to RDM_CC_x_COMMAND_RESPONSE
+  header.pdl = pdl_out;
+  // These fields should not change: tn, sub_device, and pid
+
+  // Send the response packet
+  if (response_type != RDM_RESPONSE_TYPE_NONE) {
+    const size_t response_size = dmx_write_rdm(dmx_num, &header, pd);
+    if (!dmx_send(dmx_num, response_size)) {
+      DMX_WARN("PID 0x%04x did not send a response", header.pid);
+      taskENTER_CRITICAL(DMX_SPINLOCK(dmx_num));
+      driver->flags |= DMX_FLAGS_DRIVER_BOOT_LOADER;
+      taskEXIT_CRITICAL(DMX_SPINLOCK(dmx_num));
+    } else if (response_size > 0) {
+      dmx_wait_sent(dmx_num, pdDMX_MS_TO_TICKS(23));
+      taskENTER_CRITICAL(DMX_SPINLOCK(dmx_num));
+      driver->head = -1;  // Wait for DMX break before reading data
+      dmx_uart_set_rts(driver->uart, 1);
+      taskEXIT_CRITICAL(DMX_SPINLOCK(dmx_num));
+    }
+  }
+
+  // Call the user-side callback
+  if (cb_num < driver->num_rdm_cbs && driver->rdm_cbs[cb_num].user_cb != NULL) {
+    void *context = driver->rdm_cbs[cb_num].context;
+    driver->rdm_cbs[cb_num].user_cb(dmx_num, &header, context);
+  }
+
+  // Update NVS values
+  if (driver->rdm_cbs[cb_num].non_volatile) {
+    if (!dmx_nvs_set(dmx_num, header.pid, desc->data_type, param,
+                     desc->pdl_size)) {
+      taskENTER_CRITICAL(DMX_SPINLOCK(dmx_num));
+      driver->flags |= DMX_FLAGS_DRIVER_BOOT_LOADER;
+      taskEXIT_CRITICAL(DMX_SPINLOCK(dmx_num));
+      DMX_WARN("unable to save PID 0x%04x to NVS", header.pid);
+    }
+  }
+
+  // Give the mutex back and return
+  xSemaphoreGiveRecursive(driver->mux);
+  return packet_size;
+}
+
+size_t dmx_send(dmx_port_t dmx_num, size_t size) {
+  DMX_CHECK(dmx_num < DMX_NUM_MAX, 0, "dmx_num error");
+  DMX_CHECK(dmx_driver_is_installed(dmx_num), 0, "driver is not installed");
+  DMX_CHECK(dmx_driver_is_enabled(dmx_num), 0, "driver is not enabled");
+
+  dmx_driver_t *const driver = dmx_driver[dmx_num];
+
+  // Block until the mutex can be taken
+  if (!xSemaphoreTakeRecursive(driver->mux, 0)) {
+    return 0;
+  }
+
+  // Block until the driver is done sending
+  if (!dmx_wait_sent(dmx_num, pdDMX_MS_TO_TICKS(23))) {
+    xSemaphoreGiveRecursive(driver->mux);
+    return 0;
+  }
+
+  // Determine if it is too late to send a response packet
+  int64_t elapsed = 0;
+  taskENTER_CRITICAL(DMX_SPINLOCK(dmx_num));
+  const rdm_cc_t cc = driver->data[20];
+  if (*(uint16_t *)driver->data == (RDM_SC | (RDM_SUB_SC << 8)) &&
+      (cc == RDM_CC_DISC_COMMAND_RESPONSE ||
+       cc == RDM_CC_GET_COMMAND_RESPONSE ||
+       cc == RDM_CC_SET_COMMAND_RESPONSE)) {
+    elapsed = dmx_timer_get_micros_since_boot() - driver->last_slot_ts;
+  }
+  taskEXIT_CRITICAL(DMX_SPINLOCK(dmx_num));
+  if (elapsed >= RDM_PACKET_SPACING_RESPONDER_NO_RESPONSE) {
+    xSemaphoreGiveRecursive(driver->mux);
+    return 0;
+  }
+
+  // Determine if an alarm needs to be set to wait until driver is ready
+  uint32_t timeout = 0;
+  taskENTER_CRITICAL(DMX_SPINLOCK(dmx_num));
+  if (driver->flags & DMX_FLAGS_DRIVER_SENT_LAST) {
+    if (driver->rdm_type & DMX_FLAGS_RDM_IS_DISC_UNIQUE_BRANCH) {
+      timeout = RDM_PACKET_SPACING_DISCOVERY_NO_RESPONSE;
+    } else if (driver->rdm_type & DMX_FLAGS_RDM_IS_BROADCAST) {
+      timeout = RDM_PACKET_SPACING_BROADCAST;
+    } else if (driver->rdm_type == DMX_FLAGS_RDM_IS_REQUEST) {
+      timeout = RDM_PACKET_SPACING_REQUEST_NO_RESPONSE;
+    }
+  } else if (driver->rdm_type & DMX_FLAGS_RDM_IS_VALID) {
+    timeout = RDM_PACKET_SPACING_RESPONSE;
+  }
+  elapsed = dmx_timer_get_micros_since_boot() - driver->last_slot_ts;
+  if (elapsed < timeout) {
+    dmx_timer_set_counter(driver->timer, elapsed);
+    dmx_timer_set_alarm(driver->timer, timeout, false);
+    dmx_timer_start(driver->timer);
+    driver->task_waiting = xTaskGetCurrentTaskHandle();
+  }
+  taskEXIT_CRITICAL(DMX_SPINLOCK(dmx_num));
+
+  // Block if an alarm was set
+  if (elapsed < timeout) {
+    // FIXME: clean up this section
+    bool notified = xTaskNotifyWait(0, ULONG_MAX, NULL, pdDMX_MS_TO_TICKS(23));
+    if (!notified) {
+      dmx_timer_stop(driver->timer);
+      xTaskNotifyStateClear(driver->task_waiting);
+    }
+    driver->task_waiting = NULL;
+    if (!notified) {
+      xSemaphoreGiveRecursive(driver->mux);
+      return 0;
+    }
+  }
+
+  // Turn the DMX bus around and get the send size
+  taskENTER_CRITICAL(DMX_SPINLOCK(dmx_num));
+  if (dmx_uart_get_rts(driver->uart) == 1) {
+    xTaskNotifyStateClear(xTaskGetCurrentTaskHandle());
+    dmx_uart_set_rts(driver->uart, 0);
+  }
+  taskEXIT_CRITICAL(DMX_SPINLOCK(dmx_num));
+
+  // Update the transmit size if desired
+  if (size > 0) {
+    if (size > DMX_PACKET_SIZE_MAX) {
+      size = DMX_PACKET_SIZE_MAX;
+    }
+    taskENTER_CRITICAL(DMX_SPINLOCK(dmx_num));
+    driver->tx_size = size;
+    taskEXIT_CRITICAL(DMX_SPINLOCK(dmx_num));
+  } else {
+    taskENTER_CRITICAL(DMX_SPINLOCK(dmx_num));
+    size = driver->tx_size;
+    taskEXIT_CRITICAL(DMX_SPINLOCK(dmx_num));
+  }
+
+  // Record the outgoing packet type
+  const rdm_pid_t pid = bswap16(*(uint16_t *)&driver->data[21]);
+  rdm_uid_t dest_uid;
+  rdm_uidcpy(&dest_uid, &driver->data[3]);
+  int rdm_type = 0;
+  if (*(uint16_t *)driver->data == (RDM_SC | (RDM_SUB_SC << 8))) {
+    rdm_type |= DMX_FLAGS_RDM_IS_VALID;
+    if (cc == RDM_CC_DISC_COMMAND || cc == RDM_CC_GET_COMMAND ||
+        cc == RDM_CC_SET_COMMAND) {
+      rdm_type |= DMX_FLAGS_RDM_IS_REQUEST;
+    }
+    if (rdm_uid_is_broadcast(&dest_uid)) {
+      rdm_type |= DMX_FLAGS_RDM_IS_BROADCAST;
+    }
+    if (pid == RDM_PID_DISC_UNIQUE_BRANCH) {
+      rdm_type |= DMX_FLAGS_RDM_IS_DISC_UNIQUE_BRANCH;
+    }
+  } else if (driver->data[0] == RDM_PREAMBLE ||
+             driver->data[0] == RDM_DELIMITER) {
+    rdm_type |= DMX_FLAGS_RDM_IS_VALID | DMX_FLAGS_RDM_IS_DISC_UNIQUE_BRANCH;
+  }
+  driver->rdm_type = rdm_type;
+  driver->flags |= DMX_FLAGS_DRIVER_SENT_LAST;
+  if ((rdm_type & (DMX_FLAGS_RDM_IS_VALID | DMX_FLAGS_RDM_IS_REQUEST)) ==
+      (DMX_FLAGS_RDM_IS_VALID | DMX_FLAGS_RDM_IS_REQUEST)) {
+    ++driver->tn;
+  }
+
+  // Determine if a DMX break is required and send the packet
+  if (rdm_type ==
+      (DMX_FLAGS_RDM_IS_VALID | DMX_FLAGS_RDM_IS_DISC_UNIQUE_BRANCH)) {
+    // RDM discovery responses do not send a DMX break - write immediately
+    taskENTER_CRITICAL(DMX_SPINLOCK(dmx_num));
+    driver->flags |= DMX_FLAGS_DRIVER_IS_SENDING;
+
+    size_t write_size = driver->tx_size;
+    dmx_uart_write_txfifo(driver->uart, driver->data, &write_size);
+    driver->head = write_size;
+
+    // Enable DMX write interrupts
+    dmx_uart_enable_interrupt(driver->uart, DMX_INTR_TX_ALL);
+    taskEXIT_CRITICAL(DMX_SPINLOCK(dmx_num));
+  } else {
+    // Send the packet by starting the DMX break
+    taskENTER_CRITICAL(DMX_SPINLOCK(dmx_num));
+    driver->head = 0;
+    driver->flags |=
+        (DMX_FLAGS_DRIVER_IS_IN_BREAK | DMX_FLAGS_DRIVER_IS_SENDING);
+    dmx_timer_set_counter(driver->timer, 0);
+    dmx_timer_set_alarm(driver->timer, driver->break_len, true);
+    dmx_timer_start(driver->timer);
+
+    dmx_uart_invert_tx(driver->uart, 1);
+    taskEXIT_CRITICAL(DMX_SPINLOCK(dmx_num));
+  }
+
+  // Give the mutex back
+  xSemaphoreGiveRecursive(driver->mux);
+  return size;
+}
+
+bool dmx_sniffer_enable(dmx_port_t dmx_num, int intr_pin) {
+  DMX_CHECK(dmx_num < DMX_NUM_MAX, false, "dmx_num error");
+  DMX_CHECK(dmx_sniffer_pin_is_valid(intr_pin), false, "intr_pin error");
+  DMX_CHECK(!dmx_sniffer_is_enabled(dmx_num), false,
+            "sniffer is already enabled");
+
+  dmx_driver_t *const driver = dmx_driver[dmx_num];
+
+  // Allocate the sniffer queue
+  driver->metadata_queue = xQueueCreate(1, sizeof(dmx_metadata_t));
+  DMX_CHECK(driver->metadata_queue != NULL, false,
+            "DMX sniffer queue malloc error");
+
+  // Set sniffer default values
+  driver->last_neg_edge_ts = -1;  // Negative edge hasn't been seen yet
+  driver->flags &= ~DMX_FLAGS_DRIVER_IS_IN_MAB;
+
+  // Add the GPIO interrupt handler
+  driver->gpio = dmx_gpio_init(dmx_num, driver, intr_pin);
+
+  return true;
+}
+
+bool dmx_sniffer_disable(dmx_port_t dmx_num) {
+  DMX_CHECK(dmx_num < DMX_NUM_MAX, false, "dmx_num error");
+  DMX_CHECK(dmx_sniffer_is_enabled(dmx_num), false, "sniffer is not enabled");
+
+  dmx_driver_t *const driver = dmx_driver[dmx_num];
+
+  // Disable the interrupt and remove the interrupt handler
+  dmx_gpio_deinit(driver->gpio);
+
+  // Deallocate the sniffer queue
+  vQueueDelete(driver->metadata_queue);
+  driver->metadata_queue = NULL;
+
+  return true;
+}
 
 bool DMX_ISR_ATTR dmx_driver_is_installed(dmx_port_t dmx_num) {
   return dmx_num < DMX_NUM_MAX && dmx_driver[dmx_num] != NULL;
