+++ src/rdm_utils.c
@@ -4,7 +4,6 @@
 #include <string.h>
 
 #include "dmx/config.h"
-#include "dmx/hal.h"
 #include "dmx/nvs.h"
 #include "dmx/struct.h"
 #include "endian.h"
